Edit the whole fetch into the app.js
